Move createSessionCookie and verifySessionCookie back to FirebaseAuth.
